+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.5.57 (Pending):
+* Move LocalCacheService and ReplicantRpcRequestBuilder to org.realityforge.replicant.client.json.gwt package.
 * Move generation of DataLoadCompleteEvent event from and linkage against EventBus from
   AbstractDataLoaderService to GwtDataLoaderService and move DataLoadCompleteEvent to
   org.realityforge.replicant.client.json.gwt package.